Added AuthGuard to validate user token in apis

+++ src/addresses/addresses.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, HttpException, HttpStatus } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
 import { GetUser } from 'src/auth/decorators/user.decorator';
 import { User } from 'src/users/entities/user.entity';
@@ -8,33 +8,45 @@ import { UpdateAddressDto } from './dto/update-address.dto';
 
 @Controller('addresses')
 export class AddressesController {
-  constructor(private readonly addressesService: AddressesService) {}
+  constructor(private readonly addressesService: AddressesService) { }
 
   @Post()
-  create(@Body() createAddressDto: CreateAddressDto) {
+  @UseGuards(AuthGuard('jwt'))
+  create(@GetUser() user: User, @Body() createAddressDto: CreateAddressDto) {
+    createAddressDto.user = user;
     return this.addressesService.create(createAddressDto);
   }
 
   @Get()
   @UseGuards(AuthGuard('jwt'))
   findAll(@GetUser() user: User) {
-    return this.addressesService.findAll({where: {
-      user: user.id
-    },});
+    return this.addressesService.findAll({
+      where: {
+        user: user.id
+      },
+    });
   }
 
   @Get(':id')
   findOne(@Param('id') id: string) {
-    return this.addressesService.findOne(+id,{});
+    return this.addressesService.findOne(+id, {});
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateAddressDto: UpdateAddressDto) {
-    return this.addressesService.update(+id, updateAddressDto);
+  @UseGuards(AuthGuard('jwt'))
+  async update(@GetUser() user: User, @Param('id') id: string, @Body() updateAddressDto: UpdateAddressDto) {
+    const address = await this.addressesService.findOne(+id, { "relations": ["user"] })
+    if (address.user.id === user.id)
+      return this.addressesService.update(+id, updateAddressDto);
+    throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.addressesService.remove(+id);
+  @UseGuards(AuthGuard('jwt'))
+  async remove(@GetUser() user: User, @Param('id') id: string) {
+    const address = await this.addressesService.findOne(+id, { "relations": ["user"] })
+    if (address.user.id === user.id)
+      return this.addressesService.remove(+id);
+    throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
   }
 }
