+++ src/books/books.controller.ts
@@ -1,8 +1,11 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseInterceptors, UploadedFile } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseInterceptors, UploadedFile, UseGuards, HttpException, HttpStatus, Query } from '@nestjs/common';
+import { AuthGuard } from '@nestjs/passport';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { diskStorage } from 'multer';
+import { GetUser } from 'src/auth/decorators/user.decorator';
 import { editFileName } from 'src/generics/files/edit-file-name';
 import { imageFileFilter } from 'src/generics/files/image-file-filter';
+import { User } from 'src/users/entities/user.entity';
 import { BooksService } from './books.service';
 import { CreateBookDto } from './dto/create-book.dto';
 import { UpdateBookDto } from './dto/update-book.dto';
@@ -25,29 +28,39 @@ export class BooksController {
     }),
   )
   @Post()
-  create(@UploadedFile() file: Express.Multer.File,
-  @Body() createBookDto: CreateBookDto) {
+  @UseGuards(AuthGuard('jwt'))
+  create(@GetUser() user: User, @UploadedFile() file: Express.Multer.File, @Body() createBookDto: CreateBookDto) {
     createBookDto.image = file.filename;
+    createBookDto.user = user;
     return this.booksService.create(createBookDto);
   }
 
   @Get()
-  findAll() {
-    return this.booksService.findAll({relations: ["genres"]});
+  findAll(@Query() query ) {
+    return this.booksService.findAllSearch(query["genre"]);
   }
 
   @Get(':id')
   findOne(@Param('id') id: string) {
-    return this.booksService.findOne(+id,{relations: ["genres","reviews","ratings"]});
+    return this.booksService.findOne(+id, { relations: ["genres", "reviews", "ratings"] });
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateBookDto: UpdateBookDto) {
-    return this.booksService.update(+id, updateBookDto);
+  @UseGuards(AuthGuard('jwt'))
+  async update(@GetUser() user: User, @Param('id') id: string, @Body() updateBookDto: UpdateBookDto) {
+    const book = await this.booksService.findOne(+id, { "relations": ["user"] })
+    if (book.user.id === user.id)
+      return this.booksService.update(+id, updateBookDto);
+    else throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
+
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.booksService.remove(+id);
+  @UseGuards(AuthGuard('jwt'))
+  async remove(@GetUser() user: User, @Param('id') id: string) {
+    const book = await this.booksService.findOne(+id, { "relations": ["user"] })
+    if (book.user.id === user.id)
+      return this.booksService.remove(+id);
+    else throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
   }
 }
