+++ src/ratings/ratings.controller.ts
@@ -1,14 +1,19 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, HttpException, HttpStatus } from '@nestjs/common';
 import { RatingsService } from './ratings.service';
 import { CreateRatingDto } from './dto/create-rating.dto';
 import { UpdateRatingDto } from './dto/update-rating.dto';
+import { AuthGuard } from '@nestjs/passport';
+import { GetUser } from 'src/auth/decorators/user.decorator';
+import { User } from 'src/users/entities/user.entity';
 
 @Controller('ratings')
 export class RatingsController {
-  constructor(private readonly ratingsService: RatingsService) {}
+  constructor(private readonly ratingsService: RatingsService) { }
 
   @Post()
-  create(@Body() createRatingDto: CreateRatingDto) {
+  @UseGuards(AuthGuard('jwt'))
+  create(@GetUser() user: User, @Body() createRatingDto: CreateRatingDto) {
+    createRatingDto.user = user;
     return this.ratingsService.create(createRatingDto);
   }
 
@@ -19,16 +24,24 @@ export class RatingsController {
 
   @Get(':id')
   findOne(@Param('id') id: string) {
-    return this.ratingsService.findOne(+id,{});
+    return this.ratingsService.findOne(+id, {});
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateRatingDto: UpdateRatingDto) {
-    return this.ratingsService.update(+id, updateRatingDto);
+  @UseGuards(AuthGuard('jwt'))
+  async update(@GetUser() user: User, @Param('id') id: string, @Body() updateRatingDto: UpdateRatingDto) {
+    const rating = await this.ratingsService.findOne(+id, { "relations": ["user"] })
+    if (rating.user.id === user.id)
+      return this.ratingsService.update(+id, updateRatingDto);
+    throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.ratingsService.remove(+id);
+  @UseGuards(AuthGuard('jwt'))
+  async remove(@GetUser() user: User, @Param('id') id: string) {
+    const rating = await this.ratingsService.findOne(+id, { "relations": ["user"] })
+    if (rating.user.id === user.id)
+      return this.ratingsService.remove(+id);
+    throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
   }
 }
