+++ src/users/users.controller.ts
@@ -1,7 +1,10 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, HttpException, HttpStatus } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
+import { AuthGuard } from '@nestjs/passport';
+import { GetUser } from 'src/auth/decorators/user.decorator';
+import { User } from './entities/user.entity';
 
 @Controller('users')
 export class UsersController {
@@ -14,21 +17,27 @@ export class UsersController {
   */
   @Get()
   findAll() {
-    return this.usersService.findAll({relations: ["score"]});
+    return this.usersService.findAll({ relations: ["score"] });
   }
 
   @Get(':id')
   findOne(@Param('id') id: string) {
-    return this.usersService.findOne(+id,{relations: ["score"]});
+    return this.usersService.findOne(+id, { relations: ["score"] });
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
-    return this.usersService.update(+id, updateUserDto);
+  @UseGuards(AuthGuard('jwt'))
+  update(@GetUser() user: User, @Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+    if (id === user.id + "")
+      return this.usersService.update(+id, updateUserDto);
+    throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.usersService.remove(+id);
+  @UseGuards(AuthGuard('jwt'))
+  remove(@GetUser() user: User, @Param('id') id: string) {
+    if (id === user.id + "")
+      return this.usersService.remove(+id);
+    throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
   }
 }
